# api/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: Workflow orchestration platform for financial validation
  termsOfService: http://swagger.io/terms/
  title: Flowker API
  version: 1.0.0
servers:
- url: //localhost:4021/
paths:
...
  /v1/audit-events:
    get:
      description: Retrieves a paginated list of audit log entries with optional filtering
      parameters:
      - description: Filter by event type
        in: query
        name: eventType
        schema:
          type: string
      - description: Filter by action
        in: query
        name: action
        schema:
          type: string
      - description: Filter by result
        in: query
        name: result
        schema:
          enum:
          - SUCCESS
          - FAILED
          type: string
      - description: Filter by resource type
        in: query
        name: resourceType
        schema:
          enum:
          - workflow
          - execution
          - provider_config
          type: string
      - description: Filter by resource ID
        in: query
        name: resourceId
        schema:
          format: uuid
          type: string
      - description: Start date filter (RFC3339)
        in: query
        name: dateFrom
        schema:
          format: date-time
          type: string
      - description: End date filter (RFC3339)
        in: query
        name: dateTo
        schema:
          format: date-time
          type: string
      - description: Number of items per page
        in: query
        name: limit
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        schema:
          type: string
      - description: Sort order
        in: query
        name: sortOrder
        schema:
          default: DESC
          enum:
          - ASC
          - DESC
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.AuditEntryListOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Search audit logs
      tags:
      - Audit
  /v1/audit-events/{id}:
    get:
      description: Retrieves a single audit log entry by its event ID
      parameters:
      - description: Audit event ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.AuditEntryOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Get audit entry by ID
      tags:
      - Audit
  /v1/audit-events/{id}/verify:
    get:
      description: Verifies the hash chain integrity up to the specified audit entry
      parameters:
      - description: Audit event ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.HashChainVerificationOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Verify audit hash chain
      tags:
      - Audit
...
  /v1/dashboards/executions:
    get:
      description: Retrieves an aggregated summary of executions with optional time
        range and status filters
      parameters:
      - description: Start time filter (RFC3339)
        in: query
        name: startTime
        schema:
          format: date-time
          type: string
      - description: End time filter (RFC3339)
        in: query
        name: endTime
        schema:
          format: date-time
          type: string
      - description: Status filter
        in: query
        name: status
        schema:
          enum:
          - pending
          - running
          - completed
          - failed
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.ExecutionSummaryOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Get execution summary
      tags:
      - Dashboard
  /v1/dashboards/workflows/summary:
    get:
      description: Retrieves an aggregated summary of all workflows grouped by status
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.WorkflowSummaryOutput'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Get workflow summary
      tags:
      - Dashboard
  /v1/executions:
    get:
      description: Retrieves a paginated list of workflow executions with optional
        filters
      parameters:
      - description: Filter by workflow ID
        in: query
        name: workflowId
        schema:
          format: uuid
          type: string
      - description: Filter by status
        in: query
        name: status
        schema:
          enum:
          - pending
          - running
          - completed
          - failed
          type: string
      - description: Page size (1-100)
        in: query
        name: limit
        schema:
          default: 10
          maximum: 100
          minimum: 1
          type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        schema:
          type: string
      - description: Sort field
        in: query
        name: sortBy
        schema:
          default: startedAt
          enum:
          - startedAt
          - completedAt
          type: string
      - description: Sort direction
        in: query
        name: sortOrder
        schema:
          default: DESC
          enum:
          - ASC
          - DESC
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.ExecutionListOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: List executions
      tags:
      - Executions
...
components:
  schemas:
    api.ErrorResponse:
      properties:
        code:
          example: FLK-0001
          type: string
        message:
          example: Invalid input provided
          type: string
        title:
          example: Bad Request
          type: string
      required:
      - code
      - message
      - title
      type: object
...
    github_com_LerianStudio_flowker_pkg_model.AuditActorOutput:
      example:
        ipAddress: ipAddress
        id: id
        type: type
      properties:
        id:
          type: string
        ipAddress:
          type: string
        type:
          type: string
      type: object
    github_com_LerianStudio_flowker_pkg_model.AuditEntryListOutput:
      example:
        nextCursor: nextCursor
        hasMore: true
        items:
        - actor:
            ipAddress: ipAddress
            id: id
            type: type
          result: result
          eventId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          metadata:
            key: '{}'
          resourceId: resourceId
          previousHash: previousHash
          context:
            key: '{}'
          action: action
          eventType: eventType
          hash: hash
          resourceType: resourceType
          timestamp: timestamp
        - actor:
            ipAddress: ipAddress
            id: id
            type: type
          result: result
          eventId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          metadata:
            key: '{}'
          resourceId: resourceId
          previousHash: previousHash
          context:
            key: '{}'
          action: action
          eventType: eventType
          hash: hash
          resourceType: resourceType
          timestamp: timestamp
      properties:
        hasMore:
          type: boolean
        items:
          items:
            $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.AuditEntryOutput'
          type: array
        nextCursor:
          type: string
      type: object
    github_com_LerianStudio_flowker_pkg_model.AuditEntryOutput:
      description: Immutable audit event record with hash chain integrity.
      example:
        actor:
          ipAddress: ipAddress
          id: id
          type: type
        result: result
        eventId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        metadata:
          key: '{}'
        resourceId: resourceId
        previousHash: previousHash
        context:
          key: '{}'
        action: action
        eventType: eventType
        hash: hash
        resourceType: resourceType
        timestamp: timestamp
      properties:
        action:
          type: string
        actor:
          $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.AuditActorOutput'
        context:
          additionalProperties:
            type: object
          type: object
        eventId:
          format: uuid
          type: string
        eventType:
          type: string
        hash:
          type: string
        metadata:
          additionalProperties:
            type: object
          type: object
        previousHash:
          type: string
        resourceId:
          type: string
        resourceType:
          type: string
        result:
          type: string
        timestamp:
          type: string
      type: object
...
    github_com_LerianStudio_flowker_pkg_model.ExecutionListOutput:
      example:
        nextCursor: nextCursor
        hasMore: true
        items:
        - executionId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          completedAt: completedAt
          currentStepNumber: 0
          errorMessage: errorMessage
          totalSteps: 6
          startedAt: startedAt
          workflowId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: status
        - executionId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          completedAt: completedAt
          currentStepNumber: 0
          errorMessage: errorMessage
          totalSteps: 6
          startedAt: startedAt
          workflowId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: status
      properties:
        hasMore:
          type: boolean
        items:
          items:
            $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.ExecutionStatusOutput'
          type: array
        nextCursor:
          type: string
      type: object
...
    github_com_LerianStudio_flowker_pkg_model.ExecutionStatusOutput:
      example:
        executionId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        completedAt: completedAt
        currentStepNumber: 0
        errorMessage: errorMessage
        totalSteps: 6
        startedAt: startedAt
        workflowId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        status: status
      properties:
        completedAt:
          type: string
        currentStepNumber:
          type: integer
        errorMessage:
          type: string
        executionId:
          format: uuid
          type: string
        startedAt:
          type: string
        status:
          type: string
        totalSteps:
          type: integer
        workflowId:
          format: uuid
          type: string
      type: object
    github_com_LerianStudio_flowker_pkg_model.ExecutionSummaryOutput:
      example:
        running: 2
        total: 150
        pending: 3
        completed: 120
        failed: 25
      properties:
        completed:
          example: 120
          type: integer
        failed:
          example: 25
          type: integer
        pending:
          example: 3
          type: integer
        running:
          example: 2
          type: integer
        total:
          example: 150
          type: integer
      type: object
...
    github_com_LerianStudio_flowker_pkg_model.HashChainVerificationOutput:
      description: Result of hash chain integrity verification.
      example:
        isValid: true
        totalChecked: 6
        firstInvalidId: 0
        message: message
      properties:
        firstInvalidId:
          type: integer
        isValid:
          type: boolean
        message:
          type: string
        totalChecked:
          type: integer
      type: object
...
    github_com_LerianStudio_flowker_pkg_model.StatusCountOutput:
      example:
        count: 8
        status: active
      properties:
        count:
          example: 8
          type: integer
        status:
          example: active
          type: string
      type: object
...
    github_com_LerianStudio_flowker_pkg_model.WorkflowSummaryOutput:
      example:
        total: 15
        active: 8
        byStatus:
        - count: 8
          status: active
        - count: 8
          status: active
      properties:
        active:
          example: 8
          type: integer
        byStatus:
          items:
            $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.StatusCountOutput'
          type: array
        total:
          example: 15
          type: integer
      type: object
...